pyproject.toml
[project]
name = "llm_critic"
version = "0.0.0"
description = "A validator that validates LLM responses by grading them based on provided metrics."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8.1"
dependencies = [
    "guardrails-ai>=0.4.0",
    "litellm"
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
] 

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "tests"
]

[tool.pyright]
include = ["validator"]